+++ aiosmtpd/smtp.py
@@ -97,6 +97,7 @@ class SMTP(asyncio.StreamReaderProtocol):
         self.require_starttls = tls_context and require_starttls
         self._tls_handshake_okay = True
         self._tls_protocol = None
+        self._original_transport = None
         self.session = None
         self.envelope = None
         self.transport = None
@@ -136,7 +137,7 @@ class SMTP(asyncio.StreamReaderProtocol):
             self._writer._transport = transport
             self.transport = transport
             # Do SSL certificate checking as rfc3207 part 4.1 says.
-            # Why _extra is protected attribute?
+            # Why is _extra a protected attribute?
             self.session.ssl = self._tls_protocol._extra
             handler = getattr(self.event_handler, 'handle_STARTTLS', None)
             if handler is None:
@@ -154,20 +155,27 @@ class SMTP(asyncio.StreamReaderProtocol):
 
     def connection_lost(self, error):
         log.info('%r connection lost', self.session.peer)
+        # If STARTTLS was issued, then our transport is the SSL protocol
+        # transport, and we need to close the original transport explicitly,
+        # otherwise an unexpected eof_received() will be called *after* the
+        # connection_lost().  At that point the stream reader will already be
+        # destroyed and we'll get a traceback in super().eof_received() below.
+        if self._original_transport is not None:
+            self._original_transport.close()
         super().connection_lost(error)
         self._writer.close()
         self.transport = None
+
+    def eof_received(self):
+        log.info('%r EOF received', self.session.peer)
+        self._handler_coroutine.cancel()
+        return super().eof_received()
 
     def _client_connected_cb(self, reader, writer):
         # This is redundant since we subclass StreamReaderProtocol, but I like
         # the shorter names.
         self._reader = reader
         self._writer = writer
-
-    def eof_received(self):
-        log.info('%r EOF received', self.session.peer)
-        self._handler_coroutine.cancel()
-        return super().eof_received()
 
     def _set_post_data_state(self):
         """Reset state variables to their post-DATA state."""
@@ -374,15 +382,17 @@ class SMTP(asyncio.StreamReaderProtocol):
             self,
             self.tls_context,
             None,
-            server_side=True)
-        # Reconfigure transport layer.
-        socket_transport = self.transport
-        socket_transport._protocol = self._tls_protocol
-        # Reconfigure protocol layer. Cant understand why app transport is
-        # protected property, if it MUST be used externally.
+            server_side=True,
+            call_connection_made=True)
+        # Reconfigure transport layer.  Keep a reference to the original
+        # transport so that we can close it explicitly when the connection is
+        # lost.
+        self._original_transport = self.transport
+        self._original_transport.set_protocol(self._tls_protocol)
+        # Reconfigure the protocol layer.  Why is the app transport a protected
+        # property, if it MUST be used externally?
         self.transport = self._tls_protocol._app_transport
-        # Start handshake.
-        self._tls_protocol.connection_made(socket_transport)
+        self._tls_protocol.connection_made(self._original_transport)
 
     def _strip_command_keyword(self, keyword, arg):
         keylen = len(keyword)