+++ aiosmtpd/docs/NEWS.rst
@@ -4,6 +4,8 @@
 
 1.1 (201X-XX-XX)
 ================
+* After a ``STARTTLS`` is issued, when ``connection_lost()`` is called, the
+  original transport must be closed explicitly.
 
 1.0 (2017-05-15)
 ================